fix: meta-arguments incorrectly generated...
... and simplify code a little.

Fixed problem where the meta-arguments to the generated "arrow builder"
function were just a list of identifiers instead of a list of `FnArg`.

This caused the macro to fail when any meta-arguments were present.

Secondly, the "var" argument statements e.g.

  let a = Var::new(state.clone(), #label_a)
  let b = Var::new(state.clone(), #label_b)
  ..

are now prefixed with `arg_i`, so the generated code reads like this:

  let arg_0_a = Var::new(state.clone(), #label_a)
  let arg_1_b = Var::new(state.clone(), #label_b)

This means that we're no longer introducing arbitrary variable names
into scope (they are now prefixed `arg_`), so we don't need the
`fresh_name` function anymore.